client-influx-db/requirements.txt
asyncua==1.1.0
influxdb3-python==0.4.0
python-dotenv
